Dropdown adjustments
I scoped the `details` styling to a `dropdown--basic` class. This is because I want to have a different style of `details` dropdown, and the styling for the feature-style dropdowns (currently used on https://www.vassar.edu/admission/explore/academics/) was leaking into the new dropdowns I'd designed. I also added styling for schedule dropdowns, although this isn't finished yet.